+++ core/components/minishop3/src/MiniShop3.php
@@ -20,7 +20,7 @@ use xPDO\xPDO;
 
 class MiniShop3
 {
-    public $version = '1.2.0-beta1';
+    public $version = '1.2.2-beta1';
 
     /** @var modX $modx */
     public $modx;